Replace jQuery tablesorter with minimal JS
tablesorter is total overkill for our purposes (as is jQuery itself).
Furthermore, it appears to be unmainted, and we don't want to dig into
the huge code-base, in case the need would arise.

Therefore we implement our own simplistic table sorting, based on what
we did for cmb69/wdir_xh.