Move expand/collapse buttons to own column
Otherwise the buttons wouldn't be visible if the first column was
hidden. Furthermore, we don't want to inject contents into any column,
because that might be too intrusive and could cause style issues.

We do not insert the new column as first column, but rather append it
as last column to minimize the impact with regard to style rules such
as td:nth-child().