Close #5: Button design
The nice default look of buttons breaks down as soon as we're adding a
background-image. We could solve that by adding a span with the
respective background-image, but we prefer Ludwig's great suggestion to
simply use appropriate Unicode characters. This allows us to get rid of
the images altogether, saves bandwidth and allows to customize the
triangles directly.