Close #16: Drop support for very old browsers
This allows us to get rid of most helper functions, and also to load
the script asynchronously, which can improve the perceived loading time
of the Website (well, not that much, but it seems to be best practice,
generally, and if others would do this as well, it may be considerable
an improvement).

The exclusion of unsupported browsers should be done by not shipping
the JS at all, but dynamic loading appears to counteract the browsers
preloading, and `<script type="module">` is out of the question for now
(thanks, IE).

Oh, yeah, the browser detection is done by unrelated feature inference,
which is certainly not perfect, but still way better than UA sniffing,
and appears to work quite well.